+++ lib/domain/record/components/notification_bar/notification_bar_state.freezed.dart
