+++ lib/domain/record/components/notification_bar/notification_bar_store_parameter.freezed.dart
