+++ lib/domain/record/record_page.dart
@@ -4,8 +4,8 @@ import 'package:pilll/components/organisms/pill/pill_mark.dart';
 import 'package:pilll/components/organisms/pill/pill_sheet.dart';
 import 'package:pilll/domain/initial_setting/migrate_info.dart';
 import 'package:pilll/domain/premium_trial/premium_trial_complete_modal.dart';
-import 'package:pilll/domain/record/components/notification_bar.dart';
-import 'package:pilll/domain/record/components/notification_bar_store_parameter.dart';
+import 'package:pilll/domain/record/components/notification_bar/notification_bar.dart';
+import 'package:pilll/domain/record/components/notification_bar/notification_bar_store_parameter.dart';
 import 'package:pilll/domain/record/record_page_state.dart';
 import 'package:pilll/domain/record/record_page_store.dart';
 import 'package:pilll/domain/record/record_taken_information.dart';
